projects/data_model_work_example/requirements.txt

prefect==2.14.10
snowflake-connector-python==3.10.0
sqlalchemy==2.0.32
pandas==2.2.2
great-expectations==0.18.12
pyyaml==6.0.2
